pyproject.toml
[project]
name = "personal-assistant-bot"
version = "0.1.0"
description = "Telegram personal assistant bot with expense tracking, reminders, and LLM-powered natural language parsing"
requires-python = ">=3.12"
dependencies = [
    "python-telegram-bot>=21.0",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "alembic>=1.14",
    "pydantic-settings>=2.0",
    "google-genai>=1.0",
    "apscheduler>=3.10,<4.0",
    "mcp[cli]>=1.0",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.9",
    "mypy>=1.14",
    "aiosqlite>=0.20",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
